# backend/api.yaml
openapi: '3.0.0'

info:
  version: 1.0.0
  title: Nitori API
  description: API for cpp testing

servers:
  - url: http://127.0.0.1:3000/
    description: Dev-server
  - url: http://127.0.0.1:8080/api/v1/
    description: Webpack-dev-server

paths:
  /widgets/:
    get:
      summary: Get all existing widgets
      tags:
        - Widgets
      responses:
        200:
          description: List of widgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Widget'
        403:
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create new widget
      tags:
        - Widgets
      parameters:
        - name: name
          description: Widget name
          in: query
          required: true
          schema:
            type: string
      responses:
        201:
          description: Created new widget
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'

  /widgets/{widgetId}:
    get:
      summary: Get widget
      tags:
        - Widgets
      parameters:
        - $ref: '#/components/parameters/widgetId'
      responses:
        200:
          description: Widget id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Widget'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"

  /widgets/{widgetId}/test-specs:
    get:
      summary: Get filtered list of TestSpecs
      tags:
        - TestSpecs
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: limit
          description: Limit
          in: query
          required: false
          schema:
            type: number
            default: 100
        - name: skip
          description: Skip
          in: query
          required: false
          schema:
            type: number
            default: 0
        - name: includeSources
          in: query
          required: false
          schema:
            type: boolean
        #       Filter
        - name: name
          in: query
          schema:
            type: string
        #       Sorting
        - name: sortBy
          description: Specify field to sort by
          in: query
          required: false
          schema:
            type: string
            enum:
              - timestamp
            default: timestamp
        - $ref: '#/components/parameters/orderBy'
      responses:
        200:
          description: List of TestSpecs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSpecList'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Add TestSpec
      tags:
        - TestSpecs
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: description
          in: query
          required: true
          schema:
            type: string
            default: ""
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        200:
          description: Added TestSpec
          content:
            application/json:
              schema:
                type: object
                properties:
                  compilationResult:
                    $ref: '#/components/schemas/ExecOutput'
                  testSpec:
                    $ref: '#/components/schemas/TestSpec'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"

  /widgets/{widgetId}/test-specs/{testSpecId}:
    get:
      summary: Get TestSpecs by id
      tags:
        - TestSpecs
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: testSpecId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/revision'
        - name: includeSources
          in: query
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: Specified TestSpec
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestSpec'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"
    put:
      tags:
        - TestSpecs
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: description
          in: query
          required: true
          schema:
            type: string
            default: ""
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        200:
          description: Added TestSpec
          content:
            application/json:
              schema:
                type: object
                properties:
                  compilationResult:
                    $ref: '#/components/schemas/ExecOutput'
                  testSpec:
                    $ref: '#/components/schemas/TestSpec'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"
    delete:
      summary: Delete TestSpecs by id
      tags:
        - TestSpecs
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: testSpecId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Deleted specified TestSpec
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/widgetId'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"

  /widgets/{widgetId}/test-targets:
    get:
      summary: Get list of TestTargets
      tags:
        - TestTargets
      parameters:
        - $ref: '#/components/parameters/widgetId'
        #       Filters
        - name: limit
          description: Limit
          in: query
          required: false
          schema:
            type: number
            default: 100
        - name: skip
          description: Skip
          in: query
          required: false
          schema:
            type: number
            default: 0
        - name: testSpecId
          description: Filter by TestSpec id
          in: query
          required: false
          schema:
            type: string
        - name: timestampStart
          description: Filter by TestSpec timestamp range (end)
          in: query
          required: false
          schema:
            type: number
        - name: timestampEnd
          description: Filter by TestSpec timestamp range (end)
          in: query
          required: false
          schema:
            type: number
        - name: userDataId
          description: Filter by UserData userId
          in: query
          required: false
          schema:
            type: number
        - name: userDataLogin
          description: Filter by UserData login
          in: query
          required: false
          schema:
            type: string
        - name: userDataName
          description: Filter by UserData name
          in: query
          schema:
            type: string
        - name: userDataGroupId
          description: Filter by UserData groupId
          in: query
          required: false
          schema:
            type: number
        - name: userDataGroupName
          description: Filter by UserData groupId
          in: query
          required: false
          schema:
            type: string
        #       Sorting
        - name: sortBy
          description: Specify field to sort by
          in: query
          required: false
          schema:
            type: string
            enum:
              - timestamp
            default: timestamp
        - $ref: '#/components/parameters/orderBy'
      responses:
        200:
          description: List of TestTargets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestTargetList'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Submit new TestTarget
      tags:
        - TestTargets
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: testSpecId
          description: TestSpec id to test with
          in: query
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        200:
          description: TestTarget completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestTarget'
        404:
          $ref: "#/components/responses/NotFound"
        403:
          $ref: "#/components/responses/Unauthorized"

  /widgets/{widgetId}/test-targets/{testTargetId}:
    get:
      summary: Get TestTarget info by id
      tags:
        - TestTargets
      parameters:
        - $ref: '#/components/parameters/widgetId'
        - name: testTargetId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: TestTarget info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestTarget'
        404:
          $ref: "#/components/responses/NotFound"

components:

  parameters:

    widgetId:
      name: widgetId
      description: Widget id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/widgetId'

    revision:
      name: rev
      description: Revision of specified resource
      in: query
      schema:
        type: string
        default: ''

    orderBy:
      name: orderBy
      description: Specify sorting order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc

  responses:
    NotFound:
      description: The specified resource was not found
    Unauthorized:
      description: Unauthorized

  schemas:

    widgetId:
      type: string

    Widget:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/widgetId'
        name:
          type: string

    SourceFile:
      type: object
      properties:
        data:
          type: string
        name:
          type: string

    TestSpec:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        widgetId:
          type: string
        description:
          type: string
        timestamp:
          type: number
        sourceFiles:
          type: array
          items:
            $ref: '#/components/schemas/SourceFile'

    TestSpecList:
      type: array
      items:
        $ref: '#/components/schemas/TestSpec'

    ExecOutput:
      type: object
      properties:
        exitCode:
          type: integer
        stderr:
          type: string
        stdout:
          type: string

    UserData:
      type: object
      properties:
        usedId:
          type: number
        isAdmin:
          type: boolean
        login:
          type: string
        name:
          type: string
        groupId:
          type: number
        groupName:
          type: string

    TestTarget:
      type: object
      properties:
        _id:
          type: string
        timestamp:
          type: integer
        userData:
          $ref: '#/components/schemas/UserData'
        testSpecId:
          type: string
        testSpecRev:
          type: string
        testSpec:
          $ref: '#/components/schemas/TestSpec'
        compilerResult:
          $ref: '#/components/schemas/ExecOutput'
        linkerResult:
          $ref: '#/components/schemas/ExecOutput'
        runnerResult:
          $ref: '#/components/schemas/ExecOutput'
        sourceFiles:
          type: array
          items:
            $ref: '#/components/schemas/SourceFile'

    TestTargetList:
      type: array
      items:
        $ref: '#/components/schemas/TestTarget'

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: PHPSESSID

security:
  - cookieAuth: []